Add RPC method for ntz_pool_count

+++ src/wallet/wallet2.h
@@ -979,7 +979,7 @@ namespace tools
     uint64_t import_key_images(const std::string &filename, uint64_t &spent, uint64_t &unspent);
 
     void update_pool_state(bool refreshed = false);
-    uint16_t get_ntzpool_count(bool include_unrelayed);
+    size_t get_ntzpool_count(bool include_unrelayed);
 
     void remove_obsolete_pool_txs(const std::vector<crypto::hash> &tx_hashes);
